requirements.txt
alabaster==0.7.12
astroid==2.7.2
astropy==4.3.1
autopep8==1.5.7
Babel==2.9.1
certifi==2021.5.30
charset-normalizer==2.0.4
colorama==0.4.4
docutils==0.17.1
drms==0.6.2
idna==3.2
imagesize==1.2.0
importlib-metadata==6.6.0
isort==5.9.3
Jinja2==3.0.1
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
numpy==1.21.2
packaging==21.0
pandas==1.3.2
Pillow==9.5.0
platformdirs==2.2.0
pycodestyle==2.7.0
pyerfa==2.0.0.3
Pygments==2.10.0
pylint==2.10.2
pyparsing==2.4.7
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
rstcheck==3.3.1
six==1.16.0
snowballstemmer==2.1.0
Sphinx==4.1.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
toml==0.10.2
tqdm==4.62.2
typed-ast==1.4.3
typing-extensions==3.10.0.0
urllib3==1.26.6
wrapt==1.12.1
zipp==3.15.0
